pyproject.toml
[tool.poetry]
name = "city-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
fastapi = "^0.104.1"
pydantic = "^2.5.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
requests = "^2.31.0"
httpx = "^0.25.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
